Implementation user roler

+++ app.py
@@ -56,7 +56,7 @@ def create_user():
 
     if username and password:
         hashed_password = bcrypt.hashpw(str.encode(password), bcrypt.gensalt())
-        user = User(username = username, password = hashed_password)
+        user = User(username = username, password = hashed_password, role="user")
         db.session.add(user)
         db.session.commit()
         return jsonify({"message": "Usuario cadastrado com sucesso"})
@@ -80,6 +80,10 @@ def update_user(id_user):
     data = request.json
     user = User.query.get(id_user)
 
+
+    if id_user != current_user.id and current_user.role == "user":
+        return jsonify({"message": "Operação não permitida"}), 403
+
     if user and data.get("password"):
         hashed_password = bcrypt.hashpw(str.encode(data.get("password")), bcrypt.gensalt())
         user.password = hashed_password
@@ -94,8 +98,12 @@ def update_user(id_user):
 def delete_user(id_user):
     user = User.query.get(id_user)
 
+
+    if current_user.role != "admin":
+        return jsonify({"message": "Operação não permitida"}), 403
+
     if id_user == current_user.id:
-        return jsonify({"message": "Operação não permitida"})
+        return jsonify({"message": "delete não permitido"}), 403
 
     if user:
         db.session.delete(user)
